fix: event | user | validator

+++ internal/core/service/event.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"event-planning-app/internal/core/domain"
 	"event-planning-app/internal/core/port"
 )
@@ -17,44 +18,32 @@ func NewEventService(repository port.EventRepository) port.EventService {
 
 func (e *EventService) Create(req domain.Event) (*domain.Event, error) {
 	data, err := e.repository.Create(req)
-	if err != nil {
-		return nil, err
-	}
-
-	return data, nil
+	return data, err
 }
 
 func (e *EventService) GetAll() ([]domain.Event, error) {
 	data, err := e.repository.GetAll()
-	if err != nil {
-		return nil, err
-	}
-
-	return data, nil
+	return data, err
 }
 
 func (e *EventService) GetAllByUser(id uint) ([]domain.Event, error) {
 	data, err := e.repository.GetAllByUser(id)
-	if err != nil {
-		return nil, err
-	}
-
-	return data, nil
+	return data, err
 }
 
 func (e *EventService) GetByID(id uint) (*domain.Event, error) {
 	data, err := e.repository.GetByID(id)
+	return data, err
+}
+
+func (e *EventService) Update(req domain.Event, id uint) (*domain.Event, error) {
+	event, err := e.repository.GetByID(id)
 	if err != nil {
 		return nil, err
 	}
 
-	return data, nil
-}
-
-func (e *EventService) Update(id uint, req domain.Event) (*domain.Event, error) {
-	event, err := e.repository.GetByID(id)
-	if err != nil {
-		return nil, err
+	if event.User.ID != req.User.ID {
+		return nil, errors.New("user does not have permission to perform this action")
 	}
 
 	data, err := e.repository.Update(event, req)
@@ -65,10 +54,14 @@ func (e *EventService) Update(id uint, req domain.Event) (*domain.Event, error) 
 	return data, nil
 }
 
-func (e *EventService) Delete(id uint) error {
-	event, err := e.repository.GetByID(id)
+func (e *EventService) Delete(req domain.Event) error {
+	event, err := e.repository.GetByID(req.ID)
 	if err != nil {
 		return err
+	}
+
+	if event.User.ID != req.User.ID {
+		return errors.New("user does not have permission to perform this action")
 	}
 
 	err = e.repository.Delete(event)
